requirements.txt
anyio==4.12.1
async-timeout==5.0.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
exceptiongroup==1.3.1
Flask==3.0.0
Flask-Cors==4.0.0
gtfs-realtime-bindings==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
nyct-gtfs==2.1.0
protobuf==4.25.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
schedule==1.2.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
zipp==3.23.0
